Move stripe-related functions to stripe.c; specify _GNU_SOURCE and _FILE_OFFSET_BITS for every source file; fix includes for emalloc/ecalloc

+++ src/cmd_line.c
@@ -28,7 +28,7 @@ int action_cat_info(uint32_t inode, char *path, uint64_t pos, uint64_t file_len,
 int action_cat(uint32_t inode, char *path, uint64_t pos, uint64_t file_len,
                char *data, uint64_t data_len, void **private)
 {
-    printf("%.*s", data_len, data);
+    printf("%.*s", (int)data_len, data);
     return 0;
 }
